+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Latest Changes
+  * Add timestamp to SensorData
   * Allow usage of hostname for carla::Client and resolve them to IP address
 
 ## CARLA 0.9.4